requirements.txt
asyncio
python-dotenv
langchain-core
langchain-mistralai
streamlit
xlrd
